add more error handling
