+++ src/main/java/duke/exceptions/DukeException.java
@@ -1,4 +1,13 @@
 package duke.exceptions;
 
 public class DukeException extends Exception{
+
+    public DukeException(String message) {
+        super(message);
+    }
+
+    @Override
+    public String toString() {
+        return getMessage();
+    }
 }
