--- src/main/java/duke/exceptions/DuplicateTaskException.java
@@ -1,10 +1,0 @@
-package duke.exceptions;
-
-public class DuplicateTaskException extends DukeException {
-
-    @Override
-    public String toString() {
-        return "Task already in list!";
-    }
-}
-
