--- src/main/java/duke/exceptions/InvalidCommandException.java
@@ -1,9 +1,0 @@
-package duke.exceptions;
-
-public class InvalidCommandException extends DukeException {
-
-    @Override
-    public String toString() {
-        return "You are incomprehensible. Apply yourself.";
-    }
-}
