--- src/main/java/duke/exceptions/NoBeforeException.java
@@ -1,9 +1,0 @@
-package duke.exceptions;
-
-public class NoBeforeException extends DukeException {
-
-    @Override
-    public String toString() {
-        return "Please enter the correct due date format d/mm/yyyy [HHmm]";
-    }
-}
