--- src/main/java/duke/exceptions/NoIndexException.java
@@ -1,9 +1,0 @@
-package duke.exceptions;
-
-public class NoIndexException extends DukeException {
-
-    @Override
-    public String toString() {
-        return "Please choose a task!";
-    }
-}
