--- src/main/java/duke/exceptions/OutOfRangeException.java
@@ -1,9 +1,0 @@
-package duke.exceptions;
-
-public class OutOfRangeException extends DukeException {
-
-    @Override
-    public String toString() {
-        return "No such task exists.";
-    }
-}
