+++ src/main/resources/view/MainWindow.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="610.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="duke.MainWindow" >
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="610.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="duke.ui.MainWindow" >
     <children>
         <Button fx:id="sendButton" layoutX="274.0" layoutY="575.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="25.0" prefWidth="110" text="Send" />
         <Button fx:id="helpButton" layoutX="146.0" layoutY="575.0" mnemonicParsing="false" onAction="#help" prefHeight="25.0" prefWidth="110" text="Help" />
